Music Lab: Convert MusicPlayer to TS
